backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
implicit==0.4.8
kombu==5.2.4
numpy==1.23.1
packaging==21.3
pandas==1.4.3
prompt-toolkit==3.0.30
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
scipy==1.9.0
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
psycopg2==2.9
gunicorn>=20.0.0
scikit-surprise==1.1.1